tidy, use angular email validator for API
