+++ portal-frontend/src/app/features/edit-application/other-parcels/other-parcels.component.ts
@@ -19,6 +19,7 @@ import { ApplicationDetailedDto } from '../../../services/application/applicatio
 import { ApplicationService } from '../../../services/application/application.service';
 import { ToastService } from '../../../services/toast/toast.service';
 import { formatBooleanToString } from '../../../shared/utils/boolean-helper';
+import { getLetterCombinations } from '../../../shared/utils/number-to-letter-helper';
 import { parseStringToBoolean } from '../../../shared/utils/string-helper';
 import { EditApplicationSteps } from '../edit-application.component';
 import { DeleteParcelDialogComponent } from '../parcel-details/delete-parcel/delete-parcel-dialog.component';
@@ -297,25 +298,7 @@ export class OtherParcelsComponent implements OnInit, OnDestroy {
     this.navigateToStep.emit(step);
   }
 
-  // A B C D E F G H I J K L M N O P Q R S T U V W X Y Z
-  // AA AB AC AD AE ....
   getLetterIndex(num: number) {
-    let result = '';
-    let quotient = num + 1;
-
-    while (quotient > 0) {
-      let remainder = quotient % 26;
-
-      let letter = String.fromCharCode(remainder + 64);
-      if (remainder === 0) {
-        letter = 'Z';
-        quotient--;
-      }
-
-      result = letter + result;
-      quotient = Math.floor((quotient - 1) / 26);
-    }
-
-    return result;
+    return getLetterCombinations(num);
   }
 }
