+++ services/apps/portal/src/utils/email.helper.ts
@@ -1,0 +1,1 @@
+export const emailRegex = /^[a-z0-9._%+-]+@[a-z0-9.-]+$/;
